Consume events from the users service
