+++ payments/provider/stripe/handler/subscriptions.go
@@ -1,0 +1,12 @@
+package handler
+
+import (
+	"context"
+
+	pb "github.com/micro/services/payments/provider/proto"
+)
+
+// CreateSubscription via the Stripe API, e.g. "Subscribe John Doe to Notes Gold"
+func (h *Handler) CreateSubscription(ctx context.Context, req *pb.CreateSubscriptionRequest, rsp *pb.CreateSubscriptionResponse) error {
+	return nil
+}
