+++ payments/provider/stripe/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"github.com/micro/go-micro/v2"
 	log "github.com/micro/go-micro/v2/logger"
+	"github.com/micro/go-micro/v2/server"
 
 	"github.com/micro/services/payments/provider"
 	pb "github.com/micro/services/payments/provider/proto"
@@ -22,6 +23,9 @@ func main() {
 	// Register the provider
 	h := handler.NewHandler(service)
 	pb.RegisterProviderHandler(service.Server(), h)
+
+	// Consume events from the users service
+	micro.RegisterSubscriber("go.micro.srv.users", service.Server(), h.HandleUserEvent, server.SubscriberQueue("queue.stripe"))
 
 	// Run the service
 	if err := service.Run(); err != nil {
